moved cutominput and textarea to same folder
