+++ frontend/src/components/Inputs/index.tsx
@@ -1,0 +1,2 @@
+export { CustomInput } from './Input';
+export { TextArea } from './TextArea';
