+++ frontend/src/components/PromptIOBox/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { IOBoxBar } from './IOBoxBar';
-import { TextArea } from '../TextArea';
+import { TextArea } from '../Inputs';
 import { z } from 'zod';
 
 /**
